20240930改Long

+++ FinalS1/src/main/java/tw/Final/FinalS1/repository/WishlistRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import tw.Final.FinalS1.model.WishlistModel;
 
-public interface WishlistRepository extends JpaRepository<WishlistModel, Integer>{
+public interface WishlistRepository extends JpaRepository<WishlistModel, Long>{
 
 }
